+++ tests/test_unions.py
@@ -51,6 +51,12 @@ def fun_plain_union_valid_6(arg: tuple[int, int, int] | None = None):
     pass
 
 
+def fun_plain_union_valid_7(
+    arg: int | None = Field(default_factory=lambda x: 123),
+):
+    pass
+
+
 def fun_tagged_union_valid_1(
     arg: Annotated[Model1 | Model2 | Model3, Field(discriminator="label")],
 ):
@@ -104,6 +110,14 @@ def fun_plain_union_invalid_5(arg: int | list[Optional[int]]):
     pass
 
 
+def fun_plain_union_invalid_6(arg: int | None = Field(default=123)):
+    pass
+
+
+def fun_plain_union_invalid_7(arg: int | None = Field(default_factory=lambda: 123)):
+    pass
+
+
 def fun_non_tagged_union_invalid_1(arg: Annotated[int | str, "comment"]):
     pass
 
@@ -120,6 +134,7 @@ def test_validate_function_signature():
         fun_plain_union_valid_4,
         fun_plain_union_valid_5,
         fun_plain_union_valid_6,
+        fun_plain_union_valid_7,
         fun_tagged_union_valid_1,
         fun_non_tagged_union_valid_1,
         fun_non_tagged_union_valid_2,
@@ -135,6 +150,8 @@ def test_validate_function_signature():
         fun_plain_union_invalid_3,
         fun_plain_union_invalid_4,
         fun_plain_union_invalid_5,
+        fun_plain_union_invalid_6,
+        fun_plain_union_invalid_7,
         fun_non_tagged_union_invalid_1,
         fun_non_tagged_union_invalid_2,
     ):